refactor(teacher): TeacherMatieres sous 300 lignes — éclatement complet (H9)
356 → 151 lignes. Extraction:
- composables/useTeacherMatieres.js (chargement + navigation via useRouter)
- components/teacher/MatiereCourseCard.vue (carte présentationnelle, émet navigate, CSS verbatim)
- tests: composable + carte (7 tests verts)
